requirements.txt
scrapy>=2.11.0
python-dotenv>=1.0.0
boto3>=1.34.0
scrapy-user-agents>=0.1.1
tqdm>=4.66.0
requests>=2.31.0

